+++ app/static/scripts/doomScroll.ts
@@ -10,7 +10,7 @@ export interface Activity {
     title: string;
     type: string;
     user_id: number;
-    
+    username: string;
 }
 
 }
@@ -48,6 +48,10 @@ async function addMore(){
 }
 
 function helper(n : HTMLDivElement, a: ActivityAPI.Activity){
+    const id = document.getElementById("currentUser");
+    const numId = Number(id.innerText);
+    const admin = document.getElementById("isAdmin");
+    const numAdmin = Number(admin.innerText);
     n.classList.add("col-sm-12");
     n.classList.add("col-md-6"); 
     n.classList.add("col-lg-6");
@@ -62,7 +66,7 @@ function helper(n : HTMLDivElement, a: ActivityAPI.Activity){
     b.appendChild(c);
     let d = document.createElement("div")
     d.classList.add("card-header");
-    d.innerText = a.user_id + "'s " + a.title;
+    d.innerText = a.username + "'s " + a.title;
     c.appendChild(d);
     let y = document.createElement("div");
     y.classList.add("card-body");
@@ -82,11 +86,13 @@ function helper(n : HTMLDivElement, a: ActivityAPI.Activity){
     e.id = "routePic";
     y.appendChild(e);
     }
+    if(numId === a.user_id || numAdmin === 1){
     let f = document.createElement("button");
     f.id = a.aid.toString();
     f.innerText = "Delete Activity"
     f.onclick = () => {deleteActivity(a.aid)};
     y.appendChild(f);
+    }
     return n;
 }
 async function deleteActivity(aid: string | number) {
